+++ src/components/ui/imageViewer/theme.js
@@ -1,0 +1,9 @@
+export const ImageViewerTheme = {
+    common: {
+        error: {
+            bck: 'black',
+            border: 'red',
+            text: 'white'
+        }
+    }
+};