+++ Changelog.md
@@ -1,6 +1,6 @@
 ## Changelog
 v 2.33
-  Added removal not needed twice zoomFactor
+  Added removal not needed twice zoomFactor, Added search request validate check
 
 v 2.32
   Added Gallery updates, Added GallerySearch separate component,